requirements.txt
numpy==1.26.1
pandas==2.1.2
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
requests==2.31.0
user-agents==2.2.0
tornado==6.3.3


